pyproject.toml
[project]
name = "ml3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "h5py>=3.12.1",
    "ipykernel>=6.29.5",
    "numpy>=2.2.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "scipy>=1.15.1",
    "tinygrad>=0.10.1",
    "torch-directml>=0.2.5.dev240914",
    "tqdm>=4.67.1",
]
